fix: change DeleteAccount to avoid future error

+++ src/main/java/fr/poweroff/web/controller/html/account/Admin/DeleteAccount.java
@@ -2,6 +2,7 @@ package fr.poweroff.web.controller.html.account.Admin;
 
 import fr.poweroff.web.Registries;
 import fr.poweroff.web.models.User;
+import fr.poweroff.web.util.LoginChecker;
 import org.jetbrains.annotations.NotNull;
 
 import javax.servlet.ServletException;
@@ -10,25 +11,26 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.rmi.RemoteException;
 import java.sql.SQLException;
 
 @WebServlet(name = "Delete - Other account", value = Registries.PATH_DELETE_PEOPLE)
 public class DeleteAccount extends HttpServlet {
     @Override
     public void doGet(HttpServletRequest request, @NotNull HttpServletResponse response) throws IOException, ServletException {
-        response.setContentType("text/html");
+        if (LoginChecker.performCheck(this, request, response)) {
+            String mail = request.getParameter("mail");
+            request.setAttribute("mail", mail);
 
-        String mail = request.getParameter("mail");
-        request.setAttribute("mail", mail);
+            try {
+                User user = User.getFirst(mail);
+                assert user != null;
+                user.delete();
+            } catch (SQLException e) {
+                throw new RemoteException();
+            }
 
-        try {
-            User user = User.getFirst(mail);
-            assert user != null;
-            user.delete();
-        } catch (SQLException e) {
-            e.printStackTrace();
+            response.sendRedirect(this.getServletContext().getContextPath() + Registries.PATH_INDEX);
         }
-
-        response.sendRedirect(this.getServletContext().getContextPath() + Registries.PATH_INDEX);
     }
 }
